Clip autoscroll range in TextEditorComponent
By the time that the animation frame is delivered, the requested
position could not exist anymore. This could cause the editor component
to measure a non-existent line and, as a result, throw an exception.

With this commit we will always clip the requested range right before
trying to autoscroll to it, thus preventing the above scenario from
happening.